chore: move tests to `common-datasource` crate
